requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
slack-sdk==3.21.3
requests==2.32.2
SQLAlchemy==1.4.41
Flask-Migrate==3.1.0
APScheduler==3.10.1
Flask-Script==2.0.6
gunicorn==20.1.0
nrok
